+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Change Log
 ==========
 
-Version 1.7.2-SNAPSHOT
+Version 1.7.2
 ----------------------
 * Rewrote the `subscribeBy` support and added `blockingSubscribeBy` support for consistency